refactor: Add BasePodmanAdapter
